Clean up OrderItemDto
