+++ src/Samples/Sample.AspNetCore/Extensions/CartLineExtensions.cs
@@ -2,6 +2,7 @@
 
 using Sample.AspNetCore.Models;
 using SwedbankPay.Sdk;
+using SwedbankPay.Sdk.PaymentOrders;
 
 namespace Sample.AspNetCore.Extensions
 {
