+++ src/SwedbankPay.Sdk.Infrastructure/PaymentOrders/OrderItemDto.cs
@@ -2,6 +2,32 @@
 {
     internal class OrderItemDto
     {
+        public OrderItemDto() { }
+
+        public OrderItemDto(IOrderItem item)
+        {
+            if (item == null)
+            {
+                return;
+            }
+
+            Amount = item.Amount.InLowestMonetaryUnit;
+            Class = item.Class;
+            Description = item.Description;
+            DiscountDescription = item.DiscountDescription;
+            DiscountPrice = item.DiscountPrice?.InLowestMonetaryUnit;
+            ImageUrl = item.ImageUrl;
+            ItemUrl = item.ItemUrl;
+            Name = item.Name;
+            Quantity = item.Quantity;
+            QuantityUnit = item.QuantityUnit;
+            Reference = item.Reference;
+            Type = item.Type.Value;
+            UnitPrice = item.UnitPrice.InLowestMonetaryUnit;
+            VatAmount = item.VatAmount.InLowestMonetaryUnit;
+            VatPercent = item.VatPercent;
+        }
+
         public string Reference { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
@@ -12,8 +38,8 @@
         public string DiscountDescription { get; set; }
         public decimal Quantity { get; set; }
         public string QuantityUnit { get; set; }
-        public int UnitPrice { get; set; }
-        public int DiscountPrice { get; set; }
+        public long UnitPrice { get; set; }
+        public long? DiscountPrice { get; set; }
         public int VatPercent { get; set; }
         public long Amount { get; set; }
         public long VatAmount { get; set; }
